refactor: 패키지 명 수정 및 구조 개선
