+++ src/main/java/blackjack/domain/result/GameResult.java
@@ -1,4 +1,4 @@
-package blackjack.domain.rule;
+package blackjack.domain.result;
 
 public enum GameResult {
