+++ src/test/java/blackjack/domain/ScoreTest.java
@@ -1,4 +1,4 @@
-package blackjack.domain.rule;
+package blackjack.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
