+++ src/test/java/blackjack/domain/result/BetResultsTest.java
@@ -1,4 +1,4 @@
-package blackjack.domain.rule;
+package blackjack.domain.result;
 
 import static org.assertj.core.api.Assertions.assertThat;
